ETL/api/requirements.txt
# SKALA Physical Risk AI System - API ETL Dependencies
# Python 3.10+

# Database
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

# API Client
requests>=2.28.0
urllib3>=2.0.0
